requirements.txt
# pandas==2.0.0
# pyarrow==12.0.0
# numpy==1.26.0
# scikit-learn==1.3.0

# tqdm==4.65.0

# colorlog==6.7.0

# jsonschema==4.17.0

# fastparquet==2023.4.0

# pytest==7.4.0
# pytest-cov==4.1.0
# black==23.3.0
# flake8==6.0.0
# psutil==5.9.5

fastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.13

# ML stack (Windows + Py3.11 supported)
xgboost==2.0.3
catboost==1.2.3
imbalanced-learn==0.12.2
scikit-learn==1.4.2
scipy==1.12.0
numpy==1.26.4
tldextract==5.1.2
aiohttp==3.9.3
python-whois==0.9.3
dnspython==2.6.1




# Data handling
pandas==2.2.2
pyarrow==15.0.2
joblib==1.4.2

# API helpers
python-multipart==0.0.9
groq==0.7.0
streamlit==1.32.2
